Fix for rendering fields with `when`
Example YAML:

```yaml
  config:
    - name: cluster_info
      title: Kubernetes cluster info
      items:
      - name: airgap_install
        title: Airgap cluster
        type: bool
        value: '1'
      - name: docker_registry_address
        title: Docker registry address
        type: text
        when: '{{repl ConfigOptionEquals "airgap_install" "1" }}'
```

When the `airgap_install` box was unchecked, the `Docker Registry` field would not hide.
Upon closer investigation, it appears the `Docker Registry` field was always receiving `when: "true"`.

Turns out this was because the first time we went through the "use the builder to render all the templatable fields of the config item", we were doing that to a global copy of `*api.Release`, so the `when: true` persisted across ConfigLivePost calls,
even though it should have been recomputed from the value of `airgap_install` on every LiveRender call.

- Changed APIResolver test cases to allow for multiple rounds of
  rendering
- Added a test that sends `airgap_install=1` followed by
  `airgap_install=0`, verified that this test fails to hide the field
  with `when: {{repl ConfigOptionEquals "airgap_install" "1"}}`
- Made that test pass by adding a `deepCopyConfig` at the top of
  `APIConfigResolver`