better builder
What I Did
------------

Improve how we get `templates.Builder` instances.

How I Did it
------------

- create two new methods on `templates.BuilderBuilder`,
`BaseBuilder(api.ReleaseMetadata)`, and `FullBuilder()` which takes the
release meta and the stuff needed for a making a config context.
- replace all the `getBuilder` and friends with these methods
- didn't touch the one in depGraph since its special.

How to verify it
------------

This is mostly a cleanup/refactoring chore, so verification is just
"does it still work"

Description for the Changelog
------------

Internal refactoring to consolidate usage of template functions.

![](http://www.americancustomyachts.com/images/slides/traditionbkgd.jpg)