Split full/slim docker builds
What I Did
------------

Split full/slim docker builds into a `slim` tag, that has the ship
binary and not much else, and a `full` tag, that has `kubectl`,
`terraform`, `aws-iam-authenticator`, and whatever else we might end up
wanting to add in.

How I Did it
------------

- new Dockerfile in `deploy/Dockerfile-terraform`
        - pulls from `replicated/ship:latest` by default, customizable via build tag
        - adds in all the tools we need for doing fancy terraform appliance setups
- remove `kubectl` install from `deploy/Dockerfile`
- update goreleaser to add a copy of build with `-slim` suffix, so now
for a tag `0.22.0` we create
    - `latest`
    - `slim`
    - `0`
    - `0-slim`
    - `0.22`
    - `0.22-slim`
    - `0.22-0`
    - `0.22.0-slim`

- add to stable `deploy` circle job to also build/publish a `full` tag.  I'd like to add as well `0-full`, `0.22-full`, `0.22.0-full`, etc. But want to get initial feedback on this first.

- add to unstable `deploy_unstable` circle job to also build/publish a `alpha-full` tag.

The `-full` builds get built in the circle script rather than goreleaser, because goreleaser was giving me a weird error when trying to do multiple `dockers`.

How to verify it
------------

I've pushed the results to `dexhorthy/ship:latest` and `dexhorthy/ship:full`, so you can try any terraform-powered app yaml using either of those images.

Description for the Changelog
------------

Publish an official version of the `ship` image that includes `terraform`, `kubectl`, and `aws-iam-authenticator` bundled in. The latest version will be available under the `replicatedh/ship:full` tag in docker hub. Tagged versions will be available with the `-full` suffix, e.g. `0.30.0-full`.

:ship: